apps/requirements.txt
adtk==0.6.2
matplotlib==3.3.4
nltk==3.6.1
numpy==1.20.1
pandas==1.2.4
Pillow==9.4.0
plotly==5.13.0
pycaret==2.3.10
scikit_learn==1.2.1
streamlit==1.18.1
vaderSentiment==3.3.2
